requirements.txt
beautifulsoup4==4.9.0
Bottleneck==1.3.2
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
fastai==1.0.61
fastprogress==0.2.3
future==0.18.2
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numexpr==2.7.1
numpy==1.18.4
nvidia-ml-py3==7.352.0
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0.1
urllib3==1.25.9
azure-functions